third_party/WebKit/LayoutTests/fast/sub-pixel/shadows-computed-style.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
  #text { text-shadow: 0.5px 1.5px 2.5px red; }
  #box { box-shadow: 0.5px 1.5px 2.5px 3.5px red; }
</style>

<div id="text">Text shadow</div>
<div id="box">Box shadow</div>

<script>
  description("Tests the computed style for subpixel shadows");

  var text = document.getElementById("text");
  var box = document.getElementById("box");

  var style = window.getComputedStyle(text, null);
  shouldBe("style.getPropertyValue('text-shadow')",
           "'rgb(255, 0, 0) 0.5px 1.5px 2.5px'");
  style = window.getComputedStyle(box, null);
  shouldBe("style.getPropertyValue('box-shadow')",
           "'rgb(255, 0, 0) 0.5px 1.5px 2.5px 3.5px'");
</script>
