third_party/blink/web_tests/fast/sub-pixel/shadows-computed-style-expected.txt

Tests the computed style for subpixel shadows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.getPropertyValue('text-shadow') is 'rgb(255, 0, 0) 0.5px 1.5px 2.5px'
PASS style.getPropertyValue('box-shadow') is 'rgb(255, 0, 0) 0.5px 1.5px 2.5px 3.5px'
PASS successfullyParsed is true

TEST COMPLETE

Text shadow
Box shadow
